qa tester agent/requirements.txt
langchain
langchain-openai
langchain-community
pydantic
python-dotenv

